Navbar component moved from Home to components folder and
