+++ src/components/pages/Profile/index.jsx
@@ -3,7 +3,7 @@ import { useParams, Navigate } from "react-router-dom";
 import { getProfile } from "../../../api" 
 import "./index.css";
 
-import Navbar from "../Home/Navbar";
+import Navbar from "../../Navbar";
 import ProfilePicture from "./ProfilePicture";
 import ProfileDetails from "./ProfileDetails";
 
@@ -12,8 +12,6 @@ function Profile(props) {
     const user = props.user;
     const { id } = useParams();
     const [profile, setProfile] = useState(null);
-
-    // Make get request, find user by _id, first_name + last_name, etc., render /profile/:id by response
 
     const renderProfile = async () => {
         const data = await getProfile(id);
@@ -27,11 +25,10 @@ function Profile(props) {
     if (profile) {
         return (
             <div className="profile">
+                <Navbar user={ user }/>
                 { <h1> { profile.user.first_name }'s Profile </h1> }
             </div>
         )
-    } else {
-        <Navigate to="/404" />
     }
 }
